fastapi_news_weather/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
httpx